requirements.txt
# requirements.txt

rdkit == 2023.9.4
smilespe >= 0.0.3

scipy >= 1.11.4

gym >= 0.26.2

accelerate >= 0.26.1
chardet >= 5.2.0
tokenizers
transformers >= 4.36.2

tensorflow == 2.15.0
torch >= 2.1.2
tensorboard >= 2.15.1
jaxtyping >= 0.2.25

hydra-core >= 1.3.2
matplotlib >= 3.8.2
plotnine >= 0.12.4
seaborn >= 0.13.2

toolz >= 0.12.0

ipython >= 8.20.0
ipdb >= 0.13.13
